user service full+validation

+++ src/main/java/com/example/demo/follow/followService.java
@@ -2,6 +2,7 @@ package com.example.demo.follow;
 
 import com.example.demo.user.User;
 import com.example.demo.user.UserService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,7 +20,7 @@ public class followService {
         return followRepository.findAll();
     }
 
-    public follow follow(followRequest fr){
+    public follow follow(followRequest fr,HttpServletRequest request) throws Exception {
         long user1id= Long.parseLong(fr.getUserid());
        long user2id = Long.parseLong(fr.getFollowingid());
         follow follow=new follow(user1id, user2id);
@@ -33,6 +34,10 @@ public class followService {
                 System.out.println("user: "+user1id+  " can not follow himself");
                 return null;
             }
+            long id=  userService.extractID(request);
+            if(id!=user1id){
+                throw new Exception("Wrong user-unfollow");
+            }
             System.out.println("user: "+user1id+  " is now following this user: "+user2id);
             userService.incfol(user1id,user2id);
             return followRepository.insert(follow);
@@ -40,13 +45,17 @@ public class followService {
         return null;
     }
 
-    public follow unfollow(followRequest fr ){
+    public follow unfollow(followRequest fr , HttpServletRequest request) throws Exception {
         long user1id= Long.parseLong(fr.getUserid());
         long user2id = Long.parseLong(fr.getFollowingid());
         follow follow=new follow(user1id, user2id);
         boolean check=followRepository.existsBymid(follow.getMid());
         System.out.println(check);
         if(check){
+          long id=  userService.extractID(request);
+          if(id!=user1id){
+              throw new Exception("Wrong user-unfollow");
+          }
             System.out.println("user: "+user1id+  " unfollowed this user: "+user2id);
             userService.decfol(user1id,user2id);
             return followRepository.deleteBymid(follow.getMid()).get();
